add authMiddleware to update and delete material controller

+++ controllers/materialController.js
@@ -1,4 +1,5 @@
 const { Material } = require("../models/index");
+const { User } = require("../models/index");
 const { Op } = require("sequelize");
 const {
   getPagination,
@@ -72,6 +73,14 @@ exports.createMaterials = async (req, res) => {
 // PUT /api/materials/:id
 exports.updateMaterials = async (req, res) => {
   try {
+    const user = await User.findOne({
+      where: { email: res.locals.user.email },
+    });
+
+    if(user.roleId === 3) {
+      return res.status(403).json({ message: "User not authorized." })
+    }
+
     const { title, description, materialURL, duration, image } = req.body;
     const numAffectedRows = await Material.update({  
       title, 
@@ -87,18 +96,27 @@ exports.updateMaterials = async (req, res) => {
       res.json(updatedMaterial);
     } else {
       res
-        .status(404)
-        .json({ message: `Material with id: ${req.params.id} not found` });
+      .status(404)
+      .json({ message: `Material with id: ${req.params.id} not found` });
     }
-  } catch (error) {
+    } catch (error) {
     console.error("Error updating material:", error);
     res.status(500).json({ error: "Internal server error" });
   }
 };
 
+
 // DELETE /api/materials/:id
 exports.deleteMaterials = async (req, res) => {
   try {
+    const user = await User.findOne({
+      where: { email: res.locals.user.email },
+    });
+
+    if(user.roleId === 3) {
+      return res.status(403).json({ message: "User not authorized." })
+    }
+
     const numDeleted = await Material.destroy({ where: { id: req.params.id } });
     if (numDeleted) {
       res.status(204).json({ message: "Material deleted successfully" });
@@ -112,3 +130,4 @@ exports.deleteMaterials = async (req, res) => {
     res.status(500).json({ error: "Internal server error" });
   }
 };
+
